requirements.txt
ultralytics==8.2.38
opencv-python==4.10.0.84
numpy>=1.23.0
pyyaml>=6.0.1
matplotlib>=3.8.0
rich>=13.7.0
requests>=2.31.0

# หมายเหตุ: pyrealsense2 แนะนำให้ติดตั้งผ่านสคริปต์ setup ด้วยพารามิเตอร์ -WithRealSense
# pyrealsense2
